Fixes https://freedesktop.dachary.org/show_bug.cgi?id=41300
The attachment is proposed before submitting the bug report instead of after submitting the form. There is only one submission instead of two, which is less confusing. Furthermore, the submission is protected against double click. The description field is added to the attachment so that it no longer claims to be limited to screenshots.