http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2009-March/018981.html in a nutshell : deal with it, it won't go away
The original plan was to just have the filename. Unfortunately, it turns
out that if you do that, there are certain sites that break, because they
expect the path (and they expect a Windows path, no less). This is why
Opera and IE8 return a fake path -- not because HTML5 says to do it. In
fact I made HTML5 say it because they were doing it.