https://bugs.freedesktop.org/show_bug.cgi?id=41564
Suggestion: keep the faded green outline, perhaps changing the color to a blue
to indicate it's been selected.

In the same vein, if the user selects one, then mouses over another icon, the
description should update to the newly moused-over icon, but return to the
selected icon after the mouse leaves the icon area.